refactor: 답변 검증 시, 답변이 빈 경우에 대한 검증도 객체 생성자에서 진행 및 QuestionNotAnsweredException을 domain 하위의 exception 패키지로 변경
