+++ backend/src/main/java/reviewme/review/domain/exception/QuestionNotAnsweredException.java
@@ -1,4 +1,4 @@
-package reviewme.review.service.exception;
+package reviewme.review.domain.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import reviewme.global.exception.BadRequestException;
