+++ backend/src/test/java/reviewme/review/domain/TextAnswerTest.java
@@ -1,16 +1,21 @@
 package reviewme.review.domain;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.junit.jupiter.api.Assertions.assertAll;
 
 import org.junit.jupiter.api.Test;
-import reviewme.review.service.exception.QuestionNotAnsweredException;
+import reviewme.review.domain.exception.QuestionNotAnsweredException;
 
 class TextAnswerTest {
 
     @Test
-    void 답변이_null_인_경우_예외를_발생한다() {
+    void 답변이_없는_경우_예외를_발생한다() {
         // given, when, then
-        assertThatThrownBy(() -> new TextAnswer(1L, null))
-                .isInstanceOf(QuestionNotAnsweredException.class);
+        assertAll(
+                () -> assertThatThrownBy(() -> new TextAnswer(1L, null))
+                        .isInstanceOf(QuestionNotAnsweredException.class),
+                () -> assertThatThrownBy(() -> new TextAnswer(1L, ""))
+                        .isInstanceOf(QuestionNotAnsweredException.class)
+        );
     }
 }
